--- src/main/java/ru/mail/polis/service/stakenschneider/ServiceImpl.java
@@ -1,114 +1,0 @@
-package ru.mail.polis.service.stakenschneider;
-
-import one.nio.http.HttpServer;
-import one.nio.http.HttpServerConfig;
-import one.nio.http.HttpSession;
-import one.nio.http.Param;
-import one.nio.http.Path;
-import one.nio.http.Request;
-import one.nio.http.Response;
-import one.nio.server.AcceptorConfig;
-import org.jetbrains.annotations.NotNull;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import ru.mail.polis.dao.ByteBufferConverter;
-import ru.mail.polis.dao.DAO;
-import ru.mail.polis.dao.NoSuchElementLiteException;
-import ru.mail.polis.service.Service;
-
-import java.io.IOException;
-import java.nio.ByteBuffer;
-import java.nio.charset.StandardCharsets;
-
-public class ServiceImpl extends HttpServer implements Service {
-    private static final Logger log = LoggerFactory.getLogger(ServiceImpl.class);
-    private final DAO dao;
-
-    public ServiceImpl(final int port, @NotNull final DAO dao) throws IOException {
-        super(getConfig(port));
-        this.dao = dao;
-    }
-
-    /**
-     * Standard response for successful HTTP requests.
-     *
-     * @return HTTP status code 200 (OK)
-     */
-    @Path("/v0/status")
-    public Response status() {
-        return Response.ok("OK");
-    }
-
-    /**
-     * Provide access to entities.
-     *
-     * @param id      key of entity
-     * @param request HTTP request
-     * @return response or error
-     */
-    @Path("/v0/entity")
-    public Response entity(@Param("id") final String id, @NotNull final Request request) {
-        try {
-            if (id == null || id.isEmpty()) {
-                log.info("id is null or empty");
-                return new Response(Response.BAD_REQUEST, "Id must be not null".getBytes(StandardCharsets.UTF_8));
-            }
-            final ByteBuffer key = ByteBuffer.wrap(id.getBytes(StandardCharsets.UTF_8));
-            switch (request.getMethod()) {
-                case Request.METHOD_GET:
-                    return get(key);
-
-                case Request.METHOD_PUT:
-                    return put(key, request);
-
-                case Request.METHOD_DELETE:
-                    return delete(key);
-
-                default:
-                    return new Response(Response.METHOD_NOT_ALLOWED, Response.EMPTY);
-            }
-        } catch (IOException e) {
-            log.error("Internal error", e);
-            return new Response(Response.INTERNAL_ERROR, Response.EMPTY);
-        }
-    }
-
-    private Response get(final ByteBuffer key) throws IOException {
-        try {
-            final ByteBuffer value = dao.get(key).duplicate();
-            final byte[] valueArray = ByteBufferConverter.toArray(value);
-            return Response.ok(valueArray);
-        } catch (NoSuchElementLiteException ex) {
-            log.info("Empty value: ", ex);
-            return new Response(Response.NOT_FOUND, Response.EMPTY);
-        }
-    }
-
-    private Response put(final ByteBuffer key, final Request request) throws IOException {
-        dao.upsert(key, ByteBuffer.wrap(request.getBody()));
-        return new Response(Response.CREATED, Response.EMPTY);
-    }
-
-    private Response delete(final ByteBuffer key) throws IOException {
-        dao.remove(key);
-        return new Response(Response.ACCEPTED, Response.EMPTY);
-    }
-
-    @Override
-    public void handleDefault(final Request request, final HttpSession session) throws IOException {
-        final var response = new Response(Response.BAD_REQUEST, Response.EMPTY);
-        log.warn("Can't find handler for {}", request.getPath());
-        session.sendResponse(response);
-    }
-
-    private static HttpServerConfig getConfig(final int port) {
-        if (port <= 1024 || 65536 <= port) {
-            throw new IllegalArgumentException("invalid port");
-        }
-        final AcceptorConfig acceptor = new AcceptorConfig();
-        acceptor.port = port;
-        final HttpServerConfig config = new HttpServerConfig();
-        config.acceptors = new AcceptorConfig[]{acceptor};
-        return config;
-    }
-}
